+++ planner/core/partition_pruner_test.go
@@ -15,15 +15,12 @@
 package core_test
 
 import (
-	"bytes"
 	"fmt"
-	"sort"
 	"strings"
 	"testing"
 
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/planner/core/internal"
-	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/testkit"
 	"github.com/pingcap/tidb/testkit/testdata"
 	"github.com/stretchr/testify/require"
@@ -37,7 +34,6 @@ func TestRangeColumnPartitionPruningForIn(t *testing.T) {
 	tk.MustExec("create database test_range_col_in")
 	tk.MustExec("use test_range_col_in")
 	tk.MustExec("set @@session.tidb_partition_prune_mode='static'")
-	// TODO: Test DEFAULT partition
 
 	// case in issue-26739
 	tk.MustExec(`CREATE TABLE t1 (
@@ -215,256 +211,6 @@ func TestRangeColumnPartitionPruningForInString(t *testing.T) {
 	checkColumnStringPruningTests(tests)
 	tk.MustExec(`set names utf8mb4 collate utf8mb4_unicode_ci`)
 	checkColumnStringPruningTests(tests)
-}
-
-func TestListPartitionPruner(t *testing.T) {
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("drop database if exists test_partition;")
-	tk.MustExec("create database test_partition")
-	tk.MustExec("use test_partition")
-	tk.Session().GetSessionVars().EnableClusteredIndex = variable.ClusteredIndexDefModeIntOnly
-	// TODO: Test DEFAULT partition
-	tk.MustExec(`set @@session.tidb_regard_null_as_point=false`)
-	tk.MustExec("create table t1 (id int, a int, b int                 ) partition by list (    a    ) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10,null));")
-	tk.MustExec("create table t2 (a int, id int, b int) partition by list (a*3 + b - 2*a - b) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10,null));")
-	tk.MustExec("create table t3 (b int, id int, a int) partition by list columns (a) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10,null));")
-	tk.MustExec("create table t4 (id int, a int, b int, primary key (a)) partition by list (    a    ) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10));")
-	tk.MustExec("create table t5 (a int, id int, b int, unique key (a,b)) partition by list (a*3 + b - 2*a - b) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10,null));")
-	tk.MustExec("create table t6 (b int, id int, a int, unique key (a,b)) partition by list columns (a) (partition p0 values in (1,2,3,4,5), partition p1 values in (6,7,8,9,10,null));")
-	tk.MustExec("insert into t1 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk.MustExec("insert into t2 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk.MustExec("insert into t3 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk.MustExec("insert into t4 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10)")
-	tk.MustExec("insert into t5 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk.MustExec("insert into t6 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk.MustExec(`create table t7 (a int unsigned) partition by list (a)(partition p0 values in (0),partition p1 values in (1),partition pnull values in (null),partition p2 values in (2));`)
-	tk.MustExec("insert into t7 values (null),(0),(1),(2);")
-
-	// tk2 use to compare the result with normal table.
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("drop database if exists test_partition_2;")
-	tk2.MustExec("create database test_partition_2")
-	tk2.MustExec("use test_partition_2")
-	tk2.MustExec("create table t1 (id int, a int, b int)")
-	tk2.MustExec("create table t2 (a int, id int, b int)")
-	tk2.MustExec("create table t3 (b int, id int, a int)")
-	tk2.MustExec("create table t4 (id int, a int, b int, primary key (a));")
-	tk2.MustExec("create table t5 (a int, id int, b int, unique key (a,b));")
-	tk2.MustExec("create table t6 (b int, id int, a int, unique key (a,b));")
-	tk2.MustExec("insert into t1 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk2.MustExec("insert into t2 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk2.MustExec("insert into t3 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk2.MustExec("insert into t4 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10)")
-	tk2.MustExec("insert into t5 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk2.MustExec("insert into t6 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-	tk2.MustExec(`create table t7 (a int unsigned);`)
-	tk2.MustExec("insert into t7 values (null),(0),(1),(2);")
-
-	var input []string
-	var output []struct {
-		SQL    string
-		Result []string
-		Plan   []string
-	}
-	partitionPrunerData := plannercore.GetPartitionPrunerData()
-	partitionPrunerData.GetTestCases(t, &input, &output)
-	valid := false
-	for i, tt := range input {
-		testdata.OnRecord(func() {
-			output[i].SQL = tt
-			output[i].Result = testdata.ConvertRowsToStrings(tk.MustQuery(tt).Rows())
-			output[i].Plan = testdata.ConvertRowsToStrings(tk.MustQuery("explain format = 'brief' " + tt).Rows())
-		})
-		tk.MustQuery("explain format = 'brief' " + tt).Check(testkit.Rows(output[i].Plan...))
-		result := tk.MustQuery(tt)
-		result.Check(testkit.Rows(output[i].Result...))
-		// If the query doesn't specified the partition, compare the result with normal table
-		if !strings.Contains(tt, "partition(") {
-			result.Check(tk2.MustQuery(tt).Rows())
-			valid = true
-		}
-		require.True(t, valid)
-	}
-}
-
-func TestListDefaultPartitionPruner(t *testing.T) {
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("create database default_partition")
-	tk.MustExec("use default_partition")
-	tk.MustExec("create table t (a int, b varchar(255), key (b)) partition by list (a) (partition p0 values in (0, 10), partition p1 values in (1,2,4,8,9), partition p2 values in (3,5,default))")
-	tk.MustExec(`insert into t values (-22, "-22"), (0, "0"), (2, "2"), (3,"3"), (7,"7")`)
-	tk.MustQuery(`select * from t where a = 1`).Sort().Check(testkit.Rows())
-	tk.MustQuery(`select * from t where a = 2`).Sort().Check(testkit.Rows("2 2"))
-	tk.MustQuery(`select * from t where a <= 2`).Sort().Check(testkit.Rows("-22 -22", "0 0", "2 2"))
-	tk.MustQuery(`select * from t where a = 7`).Sort().Check(testkit.Rows("7 7"))
-	tk.MustQuery(`select * from t where a > 2`).Sort().Check(testkit.Rows("3 3", "7 7"))
-	tk.MustQuery(`select * from t where b = 7`).Sort().Check(testkit.Rows("7 7"))
-	tk.MustQuery(`select * from t where a >= 6`).Sort().Check(testkit.Rows("7 7"))
-	tk.MustQuery(`explain format = 'brief' select * from t where a >= 6`).Check(testkit.Rows(
-		"TableReader 3333.33 root partition:all data:Selection",
-		"└─Selection 3333.33 cop[tikv]  ge(default_partition.t.a, 6)",
-		"  └─TableFullScan 10000.00 cop[tikv] table:t keep order:false, stats:pseudo"))
-}
-
-func TestListColumnsPartitionPruner(t *testing.T) {
-	store, clean := testkit.CreateMockStore(t)
-	defer clean()
-	tk := testkit.NewTestKit(t, store)
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("drop database if exists test_partition;")
-	tk.MustExec("create database test_partition")
-	tk.MustExec("use test_partition")
-	tk.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk.MustExec("create table t1 (id int, a int, b int) partition by list columns (b,a) (partition p0 values in ((1,1),(2,2),(3,3),(4,4),(5,5)), partition p1 values in ((6,6),(7,7),(8,8),(9,9),(10,10),(null,10)));")
-	tk.MustExec("create table t2 (id int, a int, b int) partition by list columns (id,a,b) (partition p0 values in ((1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5)), partition p1 values in ((6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)));")
-	tk.MustExec("insert into t1 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,10,null)")
-	tk.MustExec("insert into t2 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-
-	// tk1 use to test partition table with index.
-	tk1 := testkit.NewTestKit(t, store)
-	tk1.MustExec("drop database if exists test_partition_1;")
-	tk1.MustExec(`set @@session.tidb_regard_null_as_point=false`)
-	tk1.MustExec("create database test_partition_1")
-	tk1.MustExec("use test_partition_1")
-	tk1.MustExec("set @@session.tidb_enable_list_partition = ON")
-	tk1.MustExec("create table t1 (id int, a int, b int, unique key (a,b,id)) partition by list columns (b,a) (partition p0 values in ((1,1),(2,2),(3,3),(4,4),(5,5)), partition p1 values in ((6,6),(7,7),(8,8),(9,9),(10,10),(null,10)));")
-	tk1.MustExec("create table t2 (id int, a int, b int, unique key (a,b,id)) partition by list columns (id,a,b) (partition p0 values in ((1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5)), partition p1 values in ((6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)));")
-	tk1.MustExec("insert into t1 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,10,null)")
-	tk1.MustExec("insert into t2 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-
-	// tk2 use to compare the result with normal table.
-	tk2 := testkit.NewTestKit(t, store)
-	tk2.MustExec("drop database if exists test_partition_2;")
-	tk2.MustExec(`set @@session.tidb_regard_null_as_point=false`)
-	tk2.MustExec("create database test_partition_2")
-	tk2.MustExec("use test_partition_2")
-	tk2.MustExec("create table t1 (id int, a int, b int)")
-	tk2.MustExec("create table t2 (id int, a int, b int)")
-	tk2.MustExec("insert into t1 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,10,null)")
-	tk2.MustExec("insert into t2 (id,a,b) values (1,1,1),(2,2,2),(3,3,3),(4,4,4),(5,5,5),(6,6,6),(7,7,7),(8,8,8),(9,9,9),(10,10,10),(null,null,null)")
-
-	// Default RPC encoding may cause statistics explain result differ and then the test unstable.
-	tk1.MustExec("set @@tidb_enable_chunk_rpc = on")
-
-	var input []struct {
-		SQL    string
-		Pruner string
-	}
-	var output []struct {
-		SQL       string
-		Result    []string
-		Plan      []string
-		IndexPlan []string
-	}
-	partitionPrunerData := plannercore.GetPartitionPrunerData()
-	partitionPrunerData.GetTestCases(t, &input, &output)
-	valid := false
-	for i, tt := range input {
-		// Test for table without index.
-		plan := tk.MustQuery("explain format = 'brief' " + tt.SQL)
-		planTree := testdata.ConvertRowsToStrings(plan.Rows())
-		// Test for table with index.
-		indexPlan := tk1.MustQuery("explain format = 'brief' " + tt.SQL)
-		indexPlanTree := testdata.ConvertRowsToStrings(indexPlan.Rows())
-		testdata.OnRecord(func() {
-			output[i].SQL = tt.SQL
-			output[i].Result = testdata.ConvertRowsToStrings(tk.MustQuery(tt.SQL).Rows())
-			// Test for table without index.
-			output[i].Plan = planTree
-			// Test for table with index.
-			output[i].IndexPlan = indexPlanTree
-		})
-		// compare the plan.
-		plan.Check(testkit.Rows(output[i].Plan...))
-		indexPlan.Check(testkit.Rows(output[i].IndexPlan...))
-
-		// compare the pruner information.
-		checkPrunePartitionInfo(t, tt.SQL, tt.Pruner, planTree)
-		checkPrunePartitionInfo(t, tt.SQL, tt.Pruner, indexPlanTree)
-
-		// compare the result.
-		result := tk.MustQuery(tt.SQL)
-		idxResult := tk1.MustQuery(tt.SQL)
-		result.Check(idxResult.Rows())
-		result.Check(testkit.Rows(output[i].Result...))
-
-		// If the query doesn't specified the partition, compare the result with normal table
-		if !strings.Contains(tt.SQL, "partition(") {
-			result.Check(tk2.MustQuery(tt.SQL).Rows())
-			valid = true
-		}
-	}
-	require.True(t, valid)
-}
-
-func checkPrunePartitionInfo(c *testing.T, query string, infos1 string, plan []string) {
-	infos2 := getPartitionInfoFromPlan(plan)
-	comment := fmt.Sprintf("the query is: %v, the plan is:\n%v", query, strings.Join(plan, "\n"))
-	require.Equal(c, infos1, infos2, comment)
-}
-
-type testTablePartitionInfo struct {
-	Table      string
-	Partitions string
-}
-
-// getPartitionInfoFromPlan uses to extract table partition information from the plan tree string. Here is an example, the plan is like below:
-//          "Projection_7 80.00 root  test_partition.t1.id, test_partition.t1.a, test_partition.t1.b, test_partition.t2.id, test_partition.t2.a, test_partition.t2.b",
-//          "└─HashJoin_9 80.00 root  CARTESIAN inner join",
-//          "  ├─TableReader_12(Build) 8.00 root partition:p1 data:Selection_11",
-//          "  │ └─Selection_11 8.00 cop[tikv]  1, eq(test_partition.t2.b, 6), in(test_partition.t2.a, 6, 7, 8)",
-//          "  │   └─TableFullScan_10 10000.00 cop[tikv] table:t2 keep order:false, stats:pseudo",
-//          "  └─TableReader_15(Probe) 10.00 root partition:p0 data:Selection_14",
-//          "    └─Selection_14 10.00 cop[tikv]  1, eq(test_partition.t1.a, 5)",
-//          "      └─TableFullScan_13 10000.00 cop[tikv] table:t1 keep order:false, stats:pseudo"
-//
-// The return table partition info is: t1: p0; t2: p1
-func getPartitionInfoFromPlan(plan []string) string {
-	infos := make([]testTablePartitionInfo, 0, 2)
-	info := testTablePartitionInfo{}
-	for _, row := range plan {
-		partitions := getFieldValue("partition:", row)
-		if partitions != "" {
-			info.Partitions = partitions
-			continue
-		}
-		tbl := getFieldValue("table:", row)
-		if tbl != "" {
-			info.Table = tbl
-			infos = append(infos, info)
-		}
-	}
-	sort.Slice(infos, func(i, j int) bool {
-		if infos[i].Table != infos[j].Table {
-			return infos[i].Table < infos[j].Table
-		}
-		return infos[i].Partitions < infos[j].Partitions
-	})
-	buf := bytes.NewBuffer(nil)
-	for i, info := range infos {
-		if i > 0 {
-			buf.WriteString("; ")
-		}
-		buf.WriteString(fmt.Sprintf("%v: %v", info.Table, info.Partitions))
-	}
-	return buf.String()
-}
-
-func getFieldValue(prefix, row string) string {
-	if idx := strings.Index(row, prefix); idx > 0 {
-		start := idx + len(prefix)
-		end := strings.Index(row[start:], " ")
-		if end > 0 {
-			value := row[start : start+end]
-			value = strings.Trim(value, ",")
-			return value
-		}
-	}
-	return ""
 }
 
 func TestListColumnsPartitionPrunerRandom(t *testing.T) {
